+++ debian/README
@@ -1,5 +1,5 @@
-The Debian Package st2-sso-backend
-----------------------------------
+The Debian Package st2-auth-backend-sso-saml2
+---------------------------------------------
 
 Comments regarding the Package
