+++ debian/README.Debian
@@ -1,5 +1,5 @@
-st2-sso-backend for Debian
---------------------------
+st2-auth-backend-sso-saml2 for Debian
+-------------------------------------
 
 <possible notes regarding this package - if none, delete this file>
