+++ debian/README.source
@@ -1,5 +1,5 @@
-st2-sso-backend for Debian
----------------------------
+st2-auth-backend-sso-saml2 for Debian
+-------------------------------------
 
 <this file describes information about the source package, see Debian policy
 manual section 4.14. You WILL either need to modify or delete this file>
