+++ debian/control
@@ -1,4 +1,4 @@
-Source: st2-sso-backend
+Source: st2-auth-backend-sso-saml2
 Section: Python
 Priority: optional
 Maintainer: StackStorm Engineering <opsadmin@stackstorm.com>
@@ -6,8 +6,8 @@ Build-Depends: debhelper (>= 8.0.0)
 Standards-Version: 3.9.4
 Homepage: http://stackstorm.com/
 
-Package: st2-sso-backend
+Package: st2-auth-backend-sso-saml2
 Architecture: amd64
 Pre-Depends: st2
 Depends: ${misc:Depends}, st2, xmlsec1
-Description: SSO backend for StackStorm
+Description: SAML2 SSO backend for StackStorm
