+++ debian/postinst
@@ -20,7 +20,7 @@ set -e
 install_package() {
   PIP=/opt/stackstorm/st2/bin/pip
   WHEELSDIR=/opt/stackstorm/share/wheels
-  ${PIP} install --find-links ${WHEELSDIR} --no-index --quiet --upgrade st2-enterprise-sso-backend
+  ${PIP} install --find-links ${WHEELSDIR} --no-index --quiet --upgrade st2-auth-backend-sso-saml2
 }
 
 case "$1" in
