+++ st2auth_sso_saml2/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-__version__ = '3.2dev'
+__version__ = '3.3dev'
